Move IBeatmap to its own file
